requirements.txt
numpy==1.24.3
pandas==2.0.3
torch==2.0.1
scikit-learn==1.2.2
salesforce-lavis==1.0.2
